manager: use long snackbar duration for module apply prompts
- Align with upstream: give users enough time to tap the reboot action
  on uninstall, undo and reboot-to-apply snackbars

+++ app/src/main/java/me/bmax/apatch/ui/page/apm/APMMaterial.kt
@@ -832,7 +832,7 @@ private fun ModuleListMaterial(
         }
 
         val result = snackBarHost.showSnackbar(
-            message = message, actionLabel = actionLabel, duration = SnackbarDuration.Short
+            message = message, actionLabel = actionLabel, duration = SnackbarDuration.Long
         )
         if (result) {
             reboot()
@@ -863,7 +863,7 @@ private fun ModuleListMaterial(
         }
 
         val result = snackBarHost.showSnackbar(
-            message = message, actionLabel = actionLabel, duration = SnackbarDuration.Short
+            message = message, actionLabel = actionLabel, duration = SnackbarDuration.Long
         )
         if (result) {
             reboot()
@@ -992,7 +992,7 @@ private fun ModuleListMaterial(
                                         val result = snackBarHost.showSnackbar(
                                             message = rebootToApply,
                                             actionLabel = reboot,
-                                            duration = SnackbarDuration.Short
+                                            duration = SnackbarDuration.Long
                                         )
 
                                         if (result) {
